Ai, level and spawner.
Enemy ai now make there way to the player as soon as spawned, spawner added for enemies to add wave like elements to the game. Basic level laid out and white boxxed.

Likely to be one of the last if not the last update due to Games Workshops updated terms on fan made projects ala zero tolerance.